Changing Company and User dates to long to handle dates prior to Epoch. Resolve Issue #48

+++ src/Intercom/Data/Company.cs
@@ -17,10 +17,10 @@ namespace Intercom.Data
 		public string name { get; set; }
 		public Plan plan { get; set; }
 		public string company_id { get; set; }
-		public int? remote_created_at { get; set; }
-		public int? created_at { get; set; }
-		public int? updated_at { get; set; }
-		public int? last_request_at { get; set; }
+		public long? remote_created_at { get; set; }
+		public long? created_at { get; set; }
+		public long? updated_at { get; set; }
+		public long? last_request_at { get; set; }
 		public int? monthly_spend { get; set; }
 		public int? session_count { get; set; }
 		public int? user_count { get; set; }
